+++ src/debug/tiles.rs
@@ -5,8 +5,9 @@
 use crate::prelude::*;
 use bevy::{color::palettes::css::FUCHSIA, prelude::*};
 use bevy_ecs_tilemap::{
-    map::{TilemapGridSize, TilemapType},
+    map::{TilemapGridSize, TilemapTileSize, TilemapType, TilemapSize},
     tiles::TilePos,
+    prelude::TilemapAnchor,
 };
 
 /// Configuration for the [TiledDebugTilesPlugin]
@@ -61,13 +62,13 @@ fn draw_tile_infos(
     mut commands: Commands,
     config: Res<TiledDebugTilesConfig>,
     tiles_query: Query<(Entity, &Parent, &TilePos), (With<TiledMapTile>, Without<Text2d>)>,
-    layer_query: Query<(&TilemapType, &TilemapGridSize), With<TiledMapTileLayerForTileset>>,
+    layer_query: Query<(&TilemapType, &TilemapSize, &TilemapTileSize, &TilemapGridSize, &TilemapAnchor), With<TiledMapTileLayerForTileset>>,
 ) {
     for (entity, parent, tile_pos) in tiles_query.iter() {
-        let Ok((map_type, grid_size)) = layer_query.get(parent.get()) else {
+        let Ok((map_type, map_size, tile_size, grid_size, anchor)) = layer_query.get(parent.get()) else {
             continue;
         };
-        let pos = tile_pos.center_in_world(grid_size, map_type);
+        let pos = tile_pos.center_in_world(map_size, grid_size, tile_size, map_type, anchor);
         commands.entity(entity).insert((
             Text2d::new(format!("{}x{}", tile_pos.x, tile_pos.y)),
             TextColor(config.color),